display-app/requirements.txt
kivy==2.2.1
paho-mqtt==1.6.1